backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
yt-dlp>=2025.1.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
celery==5.4.0
redis==5.0.4
flower==2.0.1
ffmpeg-python==0.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
google-auth==2.30.0
requests==2.32.3
bcrypt==4.0.1
pdf2image==1.17.0
Pillow==10.4.0
pypdf>=4.0.0
